리펙토링 작업
1. 여러 곳에서 공유될 수 있는 비지니스 로직은 domain으로 위치를 변경한다.
2. Jpa & QueryDsl
3. Service에서는 모두 @Transactional을 사용하도록 하자.
4. @NoArgsConstructor 사용을 지양해야한다.

+++ app-public-api/src/main/java/co/kurrant/app/public_api/controller/client/ApplicationFormController.java
@@ -12,6 +12,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
+import java.math.BigInteger;
 
 @Tag(name = "7. Application Form ClientType")
 @RequiredArgsConstructor
@@ -33,7 +34,7 @@ public class ApplicationFormController {
 
     @Operation(summary = "아파트 스팟 개설 신청 내역", description = "아파트 스팟 개설 신청 상세 내역을 조회한다.")
     @GetMapping("/apartments/{id}")
-    public ResponseMessage getApartmentApplicationFormDetail(HttpServletRequest httpServletRequest, @PathVariable Long id) {
+    public ResponseMessage getApartmentApplicationFormDetail(HttpServletRequest httpServletRequest, @PathVariable BigInteger id) {
         return ResponseMessage.builder()
                 .message("아파트 스팟 개설 신청 내역 조회에 성공하였습니다.")
                 .data(applicationFormService.getApartmentApplicationFormDetail(commonService.getUserId(httpServletRequest), id))
@@ -42,7 +43,7 @@ public class ApplicationFormController {
 
     @Operation(summary = "아파트 스팟 개설 신청 내역 기타 내용 저장", description = "아파트 스팟 개설 신청 내역 기타 내용을 저장한다.")
     @PutMapping("/apartments/{id}/memo")
-    public ResponseMessage updateApartmentApplicationFormMemo(HttpServletRequest httpServletRequest, @PathVariable Long id, @RequestBody ApplicationFormMemoDto applicationFormMemoDto) {
+    public ResponseMessage updateApartmentApplicationFormMemo(HttpServletRequest httpServletRequest, @PathVariable BigInteger id, @RequestBody ApplicationFormMemoDto applicationFormMemoDto) {
         applicationFormService.updateApartmentApplicationFormMemo(httpServletRequest, id, applicationFormMemoDto);
         return ResponseMessage.builder()
                 .message("아파트 스팟 개설 신청 내역의 기타 내용을 업데이트 하였습니다.")
