+++ app-public-api/src/main/java/co/kurrant/app/public_api/service/ApplicationFormService.java
@@ -15,13 +15,13 @@ public interface ApplicationFormService {
     // 아파트 스팟 개설 신청
     ApplicationFormDto registerApartmentSpot(HttpServletRequest httpServletRequest, ApartmentApplicationFormRequestDto apartmentApplicationFormRequestDto);
     // 아파트 스팟 개설 신청 기타 내용 변경
-    void updateApartmentApplicationFormMemo(HttpServletRequest httpServletRequest, Long id, ApplicationFormMemoDto applicationFormMemoDto);
+    void updateApartmentApplicationFormMemo(HttpServletRequest httpServletRequest, BigInteger id, ApplicationFormMemoDto applicationFormMemoDto);
     // 기업 스팟 개설 신청
     ApplicationFormDto registerCorporationSpot(HttpServletRequest httpServletRequest, CorporationApplicationFormRequestDto corporationApplicationFormRequestDto);
     // 기업 스팟 개설 신청 기타 내용 변경
     void updateCorporationApplicationFormMemo(HttpServletRequest httpServletRequest, Long id, ApplicationFormMemoDto applicationFormMemoDto);
     // 아파트 스팟 개설 신청 내역 상세 조회
-    ApartmentApplicationFormResponseDto getApartmentApplicationFormDetail(BigInteger userId, Long id);
+    ApartmentApplicationFormResponseDto getApartmentApplicationFormDetail(BigInteger userId, BigInteger id);
     // 기업 스팟 개설 신청 내역 상세 조회
     CorporationApplicationFormResponseDto getCorporationApplicationFormDetail(BigInteger userId, Long id);
     // 스팟 신청 날짜 리스트
